django_api/requirements.txt
Django>=1.9,<1.10
django-allauth==0.25.2
django-cors-headers==1.1.0
django-environ==0.4.0
django-grappelli==2.8.1
djangorestframework==3.3.3
djangorestframework-bulk==0.2.1
django-extensions==1.6.7
django-rest-auth==0.8.1
django-timezone-field==1.3
Pillow==3.1.1
psycopg2==2.6.1
celery==3.1.23
redis==2.10.5
uWSGI==2.0.13.1
whitenoise==3.2
django-filter==0.15.0
djangorestframework-filters==0.8.0
django-celery==3.1.17